+++ packages/toolpad-app/src/server/localMode.ts
@@ -1,4 +1,3 @@
-import * as dotenv from 'dotenv';
 import * as yaml from 'yaml';
 import * as path from 'path';
 import * as fs from 'fs/promises';
@@ -14,7 +13,6 @@ import { Emitter } from '@mui/toolpad-utils/events';
 import { errorFrom } from '@mui/toolpad-utils/errors';
 import { filterValues, hasOwnProperty, mapValues } from '@mui/toolpad-utils/collections';
 import { execa } from 'execa';
-import { truncate } from '@mui/toolpad-utils/strings';
 import config from '../config';
 import * as appDom from '../appDom';
 import { migrateUp } from '../appDom/migrations';
@@ -50,8 +48,9 @@ import {
   ResponseType as AppDomRestResponseType,
 } from '../toolpadDataSources/rest/types';
 import { LocalQuery } from '../toolpadDataSources/local/types';
-import { ProjectEvents } from '../types';
+import { ProjectEvents, ToolpadProjectOptions } from '../types';
 import { Awaitable } from '../utils/types';
+import EnvManager from './EnvManager';
 
 export function getUserProjectRoot(): string {
   const { projectDir } = config;
@@ -1200,91 +1199,6 @@ async function initToolpadFolder(root: string) {
 function getCodeComponentsFingerprint(dom: appDom.AppDom) {
   const { codeComponents = [] } = appDom.getChildNodes(dom, appDom.getApp(dom));
   return codeComponents.map(({ name }) => name).join('|');
-}
-
-interface IToolpadProject {
-  options: ToolpadProjectOptions;
-  events: Emitter<ProjectEvents>;
-  getRoot(): string;
-}
-
-class EnvManager {
-  private project: IToolpadProject;
-
-  private content: string | undefined;
-
-  private values: Awaitable<Record<string, string>> | undefined;
-
-  constructor(project: IToolpadProject) {
-    this.project = project;
-    this.initWatcher();
-  }
-
-  getEnvFilePath() {
-    return path.resolve(this.project.getRoot(), '.env');
-  }
-
-  private initWatcher() {
-    if (!this.project.options.dev) {
-      return;
-    }
-
-    const envFileWatcher = chokidar.watch([this.getEnvFilePath()]);
-    envFileWatcher.on('all', async () => {
-      this.values = await this.loadEnvFile();
-      this.project.events.emit('envChanged', {});
-    });
-  }
-
-  private async parseValues() {
-    if (!this.content) {
-      return {};
-    }
-    const parsed = dotenv.parse(this.content) as any;
-
-    const envFilePath = this.getEnvFilePath();
-    // eslint-disable-next-line no-console
-    console.log(
-      `${chalk.blue('info')}  - loaded env file "${envFilePath}" with keys ${truncate(
-        Object.keys(parsed).join(', '),
-        1000,
-      )}`,
-    );
-
-    return parsed;
-  }
-
-  private async loadEnvFile() {
-    const envFilePath = this.getEnvFilePath();
-
-    try {
-      const newContent = await fs.readFile(envFilePath, { encoding: 'utf-8' });
-
-      if (newContent !== this.content) {
-        this.content = newContent;
-        return this.parseValues();
-      }
-
-      return this.values;
-    } catch (err) {
-      if (errorFrom(err).code !== 'ENOENT') {
-        throw err;
-      }
-    }
-
-    return {};
-  }
-
-  async getValues() {
-    if (!this.values) {
-      this.values = this.loadEnvFile();
-    }
-    return this.values;
-  }
-}
-
-interface ToolpadProjectOptions {
-  dev: boolean;
 }
 
 class ToolpadProject {
